requirements.txt
selenium
pytest
webdriver-manager
pyyaml
